+++ CHANGELOG.md
@@ -7,13 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support for Sphinx style reST docs to doc parse.
-- Optional `name` argument to [tanchan.docs_parse.as_slash_command][] which
+- Optional `name` argument to [tanchan.doc_parse.as_slash_command][] which
   allows overriding the command's name.
 
 ### Fixed
 - [tanchan.doc_parse.as_slash_command][] no-longer errors when the callback's
   docstring is just the description and `doc_style` is [None][].
-- [tanjun.doc_parse.with_annotated_args][] now allows [None][] to be explicitly
+- [tanchan.doc_parse.with_annotated_args][] now allows [None][] to be explicitly
   passed to `doc_style` typing wise.
 
 ## [0.1.0] - 2022-12-02